+++ guillotina_elasticsearch/interfaces.py
@@ -1,0 +1,13 @@
+from guillotina.interfaces import ICatalogUtility
+
+
+try:
+    from guillotina.async_util import IAsyncUtility
+except ImportError:
+    from guillotina.async import IAsyncUtility
+
+
+DOC_TYPE = 'doc'
+
+class IElasticSearchUtility(ICatalogUtility, IAsyncUtility):
+    pass