+++ guillotina_elasticsearch/schema.py
@@ -45,11 +45,10 @@ def get_mappings():
     from guillotina import app_settings
     mapping_overrides = app_settings.get('elasticsearch', {}).get('mapping_overrides', {})
     # Mapping calculated from schemas
-    global_mappings = {}
+    mappings = {}
     base_type_overrides = mapping_overrides.get('*', {})
     for name, _ in get_utilities_for(IResourceFactory):
         # For each type
-        mappings = {}
         type_overrides = base_type_overrides.copy()
         type_overrides.update(mapping_overrides.get(name, {}))
         for field_name, catalog_info in get_index_fields(name).items():
@@ -63,11 +62,10 @@ def get_mappings():
             if index_name in type_overrides:
                 field_mapping = type_overrides[index_name]
             mappings[index_name] = field_mapping
-        global_mappings[name] = {
-            'properties': mappings,
-            'dynamic': False,
-            '_all': {
-                'enabled': False
-            }
+    return {
+        'properties': mappings,
+        'dynamic': False,
+        '_all': {
+            'enabled': False
         }
-    return global_mappings
+    }