+++ guillotina_elasticsearch/utils.py
@@ -1,4 +1,4 @@
-from aioes import exception
+from aioelasticsearch import exceptions
 
 import logging
 
@@ -16,10 +16,10 @@ noop_response = NoopResponse()
 async def safe_es_call(func, *args, **kwargs):
     try:
         return await func(*args, **kwargs)
-    except exception.ConnectionError:
+    except exceptions.ConnectionError:
         logger.warning('elasticsearch not installed', exc_info=True)
-    except (exception.RequestError, exception.NotFoundError,
+    except (exceptions.RequestError, exceptions.NotFoundError,
             RuntimeError):
         pass
-    except exception.TransportError as e:
+    except exceptions.TransportError as e:
         logger.warning('Transport Error', exc_info=e)